refactor: remove v3 directory
