+++ src/Converter/components/MediaType.ts
@@ -1,6 +1,6 @@
 import ts from "typescript";
 
-import { Factory } from "../../../CodeGenerator";
+import { Factory } from "../../CodeGenerator";
 import * as ConverterContext from "../ConverterContext";
 import * as ToTypeNode from "../toTypeNode";
 import { OpenApi } from "../types";
