+++ src/Converter/components/Operation.ts
@@ -3,7 +3,7 @@ import * as path from "path";
 
 import ts from "typescript";
 
-import { Factory } from "../../../CodeGenerator";
+import { Factory } from "../../CodeGenerator";
 import * as ConverterContext from "../ConverterContext";
 import * as Guard from "../Guard";
 import * as Name from "../Name";
