+++ src/Converter/components/RequestBodies.ts
@@ -1,6 +1,6 @@
 import * as path from "path";
 
-import { Factory } from "../../../CodeGenerator";
+import { Factory } from "../../CodeGenerator";
 import * as ConverterContext from "../ConverterContext";
 import * as Guard from "../Guard";
 import * as Name from "../Name";
