+++ src/Converter/components/Response.ts
@@ -1,6 +1,6 @@
 import * as path from "path";
 
-import { Factory } from "../../../CodeGenerator";
+import { Factory } from "../../CodeGenerator";
 import * as ConverterContext from "../ConverterContext";
 import * as Name from "../Name";
 import { Store } from "../store";
