+++ src/Converter/store/Operation.ts
@@ -1,6 +1,6 @@
 import { EOL } from "os";
 
-import { DevelopmentError } from "../../../Exception";
+import { DevelopmentError } from "../../Exception";
 import { OpenApi } from "../types";
 
 const httpMethodList = ["get", "put", "post", "delete", "options", "head", "patch", "trace"] as const;
