+++ src/Converter/store/Store.ts
@@ -4,8 +4,8 @@ import { Tree } from "@himenon/path-oriented-data-structure";
 import Dot from "dot-prop";
 import ts from "typescript";
 
-import { Factory } from "../../../CodeGenerator";
-import { UnSupportError } from "../../../Exception";
+import { Factory } from "../../CodeGenerator";
+import { UnSupportError } from "../../Exception";
 import { OpenApi } from "../types";
 import * as Def from "./Definition";
 import * as Operation from "./Operation";
