+++ src/Converter/store/structure/InterfaceNode.ts
