+++ src/Converter/store/structure/TypeAliasNode.ts
