+++ src/Converter/types/OpenApiSchemaV3.ts
