+++ src/DefaultCodeTemplate/ApiClientClass/MethodBody/PathParameter.ts
@@ -1,7 +1,7 @@
 import ts from "typescript";
 
 import { Factory } from "../../../CodeGenerator";
-import { PickedParameter } from "../../../Converter/v3";
+import { PickedParameter } from "../../../Converter";
 import * as Utils from "../../utils";
 
 export const isPathParameter = (params: any): params is PickedParameter => {
