+++ src/DefaultCodeTemplate/ApiClientClass/MethodBody/index.ts
@@ -1,7 +1,7 @@
 import ts from "typescript";
 
 import { Factory } from "../../../CodeGenerator";
-import { CodeGeneratorParams } from "../../../Converter/v3";
+import { CodeGeneratorParams } from "../../../Converter";
 import * as Utils from "../../utils";
 import * as CallRequest from "./CallRequest";
 import * as HeaderParameter from "./HeaderParameter";
